Add Geometry module
