+++ hvega/src/Graphics/Vega/VegaLite/Core.hs
@@ -17,21 +17,7 @@ not directly exported to the user.
 -}
 
 module Graphics.Vega.VegaLite.Core
-       ( geometry
-       , geoFeatureCollection
-       , geometryCollection
-       , Geometry(..)
-
-       , sphere
-       , graticule
-       , GraticuleProperty(..)
-
-       , transform
-
-       , projection
-       , ProjectionProperty(..)
-       , Projection(..)
-       , ClipRect(..)
+       ( transform
 
        , aggregate
        , joinAggregate
@@ -248,7 +234,6 @@ module Graphics.Vega.VegaLite.Core
        , stackOffset
        , titleConfigSpec
        , autosizeProperty
-       , projectionProperty
        , paddingSpec
        , header_
        , mprops_
@@ -262,8 +247,6 @@ import Prelude hiding (filter, lookup, repeat)
 import qualified Data.Aeson as A
 import qualified Data.Text as T
 import qualified Data.Vector as V
-
-import Control.Arrow (second)
 
 -- Aeson's Value type conflicts with the Number type
 import Data.Aeson (Value, decode, encode, object, toJSON, (.=))
@@ -321,9 +304,6 @@ clamped xmin xmax x = max xmin (min xmax x)
 
 aggregate_ :: Operation -> LabelledSpec
 aggregate_ op = "aggregate" .= operationSpec op
-
-type_ :: T.Text -> LabelledSpec
-type_ t = "type" .= t
 
 field_ :: T.Text -> LabelledSpec
 field_ f = "field" .= f
@@ -417,48 +397,6 @@ fromT = toJSON
 
 fromF :: Double -> VLSpec
 fromF = toJSON
-
-
-{-|
-
-Specifies a list of geo features to be used in a geoShape specification.
-Each feature object in this collection can be created with the 'geometry'
-function.
-
-@
-'geoFeatureCollection'
-    [ 'geometry' ('GeoPolygon' [ [ ( -3, 59 ), ( -3, 52 ), ( 4, 52 ), ( -3, 59 ) ] ])
-        [ ( "myRegionName", 'Str' "Northern region" ) ]
-    , 'geometry' ('GeoPolygon' [ [ ( -3, 52 ), ( 4, 52 ), ( 4, 45 ), ( -3, 52 ) ] ])
-        [ ( "myRegionName", 'Str' "Southern region" ) ]
-    ]
-@
--}
-geoFeatureCollection :: [VLSpec] -> VLSpec
-geoFeatureCollection geoms =
-  object [ type_ "FeatureCollection"
-         , "features" .=  geoms
-         ]
-
-
-{-|
-
-Specifies a list of geometry objects to be used in a geoShape specification.
-Each geometry object in this collection can be created with the 'geometry'
-function.
-
-@
-'geometryCollection'
-    [ 'geometry' ('GeoPolygon' [ [ ( -3, 59 ), ( 4, 59 ), ( 4, 52 ), ( -3, 59 ) ] ]) []
-    , 'geometry' ('GeoPoint' -3.5 55.5) []
-    ]
-@
--}
-geometryCollection :: [VLSpec] -> VLSpec
-geometryCollection geoms =
-  object [ type_ "GeometryCollection"
-         , "geometries" .= geoms
-         ]
 
 
 {-|
@@ -1126,7 +1064,7 @@ dimensions on a plane and @X2@ and @Y2@ represent secondary axis dimensions wher
 two scales are overlaid in the same space. Geographic positions represented by
 longitude and latiutude values are identified with @Longitude@, @Latitude@ and
 their respective secondary equivalents. Such geographic position channels are
-subject to a map projection (set using 'projection') before being placed graphically.
+subject to a map projection (set using 'Graphics.Vega.VegaLite.projection') before being placed graphically.
 
 -}
 data Position
@@ -3717,92 +3655,6 @@ pointMarkerSpec (PMMarker []) = toJSON True
 pointMarkerSpec (PMMarker mps) = object (map markProperty mps)
 
 
-{-|
-
-Types of geographic map projection. These are based on a subset of those provided
-by the <https://github.com/d3/d3-geo d3-geo library>. For details of available
-projections see the
-<https://vega.github.io/vega-lite/docs/projection.html#projection-types Vega-Lite documentation>.
--}
-
--- based on schema 3.3.0 #/definitions/ProjectionType
-
-data Projection
-    = Albers
-      -- ^ An Albers equal-area conic map projection.
-    | AlbersUsa
-      -- ^ An Albers USA map projection that combines continental USA with
-      --   Alaska and Hawaii. Unlike other projection types, this remains
-      --   unaffected by 'PrRotate'.
-    | AzimuthalEqualArea
-      -- ^ An azimuthal equal area map projection.
-    | AzimuthalEquidistant
-      -- ^ An azimuthal equidistant map projection.
-    | ConicConformal
-      -- ^ A conformal conic map projection.
-    | ConicEqualArea
-      -- ^ An equal area conic map projection.
-    | ConicEquidistant
-      -- ^ An equidistant conic map projection.
-    | Custom T.Text
-      -- ^ Specify the name of the custom D3 prohection to use. See the
-      --   <https://vega.github.io/vega/docs/projections/#register Vega API>
-      --   for more information.
-      --
-      --   An example: @Custom "winkle3"@
-    | Equirectangular
-      -- ^ An equirectangular map projection that maps longitude to x and latitude to y.
-      --   While showing less area distortion towards the poles than the default 'Mercator'
-      --   projection, it is neither equal-area nor conformal.
-    | Gnomonic
-      -- ^ A gnomonic map projection.
-    | Identity
-      -- ^ The identiy projection. This can be combined with 'PrReflectX' and
-      --   'PrReflectY' in the list of projection properties.
-      --
-      --   @since 0.4.0.0
-    | Mercator
-      -- ^ A Mercator map projection. This is the default projection of longitude, latitude
-      --   values if no projection is set explicitly. It preserves shape (local angle) and
-      --   lines of equal angular bearing remain parallel straight lines. The area is
-      --   /significantly/ enlarged towards the poles.
-    | Orthographic
-      -- ^ An orthographic map projection.
-    | Stereographic
-      -- ^ A stereographic map projection.
-    | TransverseMercator
-      -- ^ A transverse Mercator map projection.
-
-
-projectionLabel :: Projection -> T.Text
-projectionLabel Albers = "albers"
-projectionLabel AlbersUsa = "albersUsa"
-projectionLabel AzimuthalEqualArea = "azimuthalEqualArea"
-projectionLabel AzimuthalEquidistant = "azimuthalEquidistant"
-projectionLabel ConicConformal = "conicConformal"
-projectionLabel ConicEqualArea = "conicEqualarea"
-projectionLabel ConicEquidistant = "conicEquidistant"
-projectionLabel (Custom pName) = pName
-projectionLabel Equirectangular = "equirectangular"
-projectionLabel Gnomonic = "gnomonic"
-projectionLabel Identity = "identity"
-projectionLabel Mercator = "mercator"
-projectionLabel Orthographic = "orthographic"
-projectionLabel Stereographic = "stereographic"
-projectionLabel TransverseMercator = "transverseMercator"
-
-
--- | Specifies a clipping rectangle for defining
---   the clip extent of a map projection.
-
-data ClipRect
-    = NoClip
-      -- ^ No clipping it to be applied.
-    | LTRB Double Double Double Double
-      -- ^ The left, top, right, and bottom extents, in pixels,
-      --   of a rectangular clip.
-
-
 -- | Specifies the alignment of compositions. It is used with:
 --   'align', 'alignRC', 'LeGridAlign', and 'LGridAlign'.
 --
@@ -3823,126 +3675,6 @@ compositionAlignmentSpec :: CompositionAlignment -> VLSpec
 compositionAlignmentSpec CANone = "none"
 compositionAlignmentSpec CAEach = "each"
 compositionAlignmentSpec CAAll = "all"
-
-
-{-|
-
-Properties for customising a geospatial projection that converts longitude,latitude
-pairs into planar @(x,y)@ coordinate pairs for rendering and query. For details see the
-<https://vega.github.io/vega-lite/docs/projection.html Vega-Lite documentation>.
-
-This type has been changed in the @0.4.0.0@ release so that all constructors
-start with @Pr@ rather than @P@ (and so provide some differentiation to the
-'PositionChannel' constructors).
-
--}
-
--- based on schema 3.3.0 #/definitions/Projection
-
-data ProjectionProperty
-    = PrType Projection
-      -- ^ The type of the map projection.
-    | PrClipAngle (Maybe Double)
-      -- ^ The clipping circle angle in degrees. A value of @Nothing@ will switch to
-      --   antimeridian cutting rather than small-circle clipping.
-    | PrClipExtent ClipRect
-      -- ^ Projection’s viewport clip extent to the specified bounds in pixels.
-    | PrCenter Double Double
-      -- ^ Projection’s center as longitude and latitude in degrees.
-    | PrScale Double
-      -- ^ The projection's zoom scale, which if set, overrides automatic scaling of a
-      --   geo feature to fit within the viewing area.
-      --
-      --   @since 0.4.0.0
-    | PrTranslate Double Double
-      -- ^ A projection’s panning translation, which if set, overrides automatic positioning
-      --   of a geo feature to fit within the viewing area
-      --
-      --   Note that the prefix is @Pr@ and not @P@, to match the Elm API.
-      --
-      --   @since 0.4.0.0
-    | PrRotate Double Double Double
-      -- ^ A projection’s three-axis rotation angle. The order is @lambda@ @phi@ @gamma@,
-      --   and specifies the rotation angles in degrees about each spherical axis.
-    | PrPrecision Double
-      -- ^ Threshold for the projection’s adaptive resampling in pixels, and corresponds to the
-      --   Douglas–Peucker distance. If precision is not specified, the projection’s current
-      --   resampling precision of 0.707 is used.
-      --
-      --   Version 3.3.0 of the Vega-Lite spec claims this should be output as a string,
-      --   but it is written out as a number since the
-      --   [spec is in error](https://github.com/vega/vega-lite/issues/5190).
-    | PrReflectX Bool
-      -- ^ Reflect the x-coordinates after performing an identity projection. This
-      --   creates a left-right mirror image of the geoshape marks when subject to an
-      --   identity projection with 'Identity'.
-      --
-      -- @since 0.4.0.0
-    | PrReflectY Bool
-      -- ^ Reflect the y-coordinates after performing an identity projection. This
-      --   creates a left-right mirror image of the geoshape marks when subject to an
-      --   identity projection with 'Identity'.
-      --
-      -- @since 0.4.0.0
-    | PrCoefficient Double
-      -- ^ The @Hammer@ map projection coefficient.
-    | PrDistance Double
-      -- ^ The @Satellite@ map projection distance.
-    | PrFraction Double
-      -- ^ The @Bottomley@ map projection fraction.
-    | PrLobes Int
-      -- ^ Number of lobes in lobed map projections such as the @Berghaus star@.
-    | PrParallel Double
-      -- ^ Parallel for map projections such as the @Armadillo@.
-    | PrRadius Double
-      -- ^ Radius value for map projections such as the @Gingery@.
-    | PrRatio Double
-      -- ^ Ratio value for map projections such as the @Hill@.
-    | PrSpacing Double
-      -- ^ Spacing value for map projections such as the @Lagrange@.
-    | PrTilt Double
-      -- ^ @Satellite@ map projection tilt.
-
-
-projectionProperty :: ProjectionProperty -> LabelledSpec
-projectionProperty (PrType proj) = "type" .= projectionLabel proj
-projectionProperty (PrClipAngle numOrNull) = "clipAngle" .= maybe A.Null toJSON numOrNull
-projectionProperty (PrClipExtent rClip) =
-  ("clipExtent", case rClip of
-    NoClip -> A.Null
-    LTRB l t r b -> toJSON (map toJSON [l, t, r, b])
-  )
-projectionProperty (PrCenter lon lat) = "center" .= [lon, lat]
-projectionProperty (PrScale sc) = "scale" .= sc
-projectionProperty (PrTranslate tx ty) = "translate" .= [tx, ty]
-projectionProperty (PrRotate lambda phi gamma) = "rotate" .= [lambda, phi, gamma]
-projectionProperty (PrPrecision pr) = "precision" .= pr  -- the 3.3.0 spec says this is a string, but that's wrong,  See https://github.com/vega/vega-lite/issues/5190
-projectionProperty (PrReflectX b) = "reflectX" .= b
-projectionProperty (PrReflectY b) = "reflectY" .= b
-projectionProperty (PrCoefficient x) = "coefficient" .= x
-projectionProperty (PrDistance x) = "distance" .= x
-projectionProperty (PrFraction x) = "fraction" .= x
-projectionProperty (PrLobes n) = "lobes" .= n
-projectionProperty (PrParallel x) = "parallel" .= x
-projectionProperty (PrRadius x) = "radius" .= x
-projectionProperty (PrRatio x) = "ratio" .= x
-projectionProperty (PrSpacing x) = "spacing" .= x
-projectionProperty (PrTilt x) = "tilt" .= x
-
-
-{-|
-
-Sets the cartographic projection used for geospatial coordinates. A projection
-defines the mapping from @(longitude,latitude)@ to an @(x,y)@ plane used for rendering.
-This is useful when using the 'Geoshape' mark. For further details see the
-<https://vega.github.io/vega-lite/docs/projection.html Vega-Lite documentation>.
-
-@
-'projection' [ 'PrType' 'Orthographic', 'PrRotate' (-40) 0 0 ]
-@
--}
-projection :: [ProjectionProperty] -> PropertySpec
-projection pProps = (VLProjection, object (map projectionProperty pProps))
 
 
 {-|
@@ -4783,170 +4515,6 @@ vale to accept and the second the inclusive maximum.
 data FilterRange
     = NumberRange Double Double
     | DateRange [DateTime] [DateTime]
-
-
-{-|
-
-Specifies the type and content of geometry specifications for programatically
-creating GeoShapes. These can be mapped to the
-<https://tools.ietf.org/html/rfc7946#section-3.1 GeoJson geometry object types>
-where the pluralised type names refer to their @Multi@ prefixed equivalent in the
-GeoJSON specification.
--}
-data Geometry
-    = GeoPoint Double Double
-    -- ^ The GeoJson geometry @point@ type.
-    | GeoPoints [(Double, Double)]
-    -- ^ The GeoJson geometry @multi-point@ type.
-    | GeoLine [(Double, Double)]
-    -- ^ The GeoJson geometry @line@ type.
-    | GeoLines [[(Double, Double)]]
-    -- ^ The GeoJson geometry @multi-line@ type.
-    | GeoPolygon [[(Double, Double)]]
-    -- ^ The GeoJson geometry @polygon@ type.
-    | GeoPolygons [[[(Double, Double)]]]
-    -- ^ The GeoJson geometry @multi-polygon@ type.
-
-
-{-|
-
-Specifies a geometric object to be used in a geoShape specification. The first
-parameter is the geometric type, the second an optional list of properties to be
-associated with the object.
-
-@
-'geometry' ('GeoPolygon' [ [ ( -3, 59 ), ( 4, 59 ), ( 4, 52 ), ( -3, 59 ) ] ]) []
-@
--}
-geometry :: Geometry -> [(T.Text, DataValue)] -> VLSpec
-geometry gType properties =
-  object ([ ("type", fromT "Feature")
-          , ("geometry", geometryTypeSpec gType) ]
-          <> if null properties
-             then []
-             else [("properties",
-                    object (map (second dataValueSpec) properties))]
-         )
-
-
-geometryTypeSpec :: Geometry -> VLSpec
-geometryTypeSpec gType =
-  let toCoords :: [(Double, Double)] -> VLSpec
-      toCoords = toJSON -- rely on Aeson converting a pair to a 2-element list
-
-      toCoordList :: [[(Double, Double)]] -> VLSpec
-      toCoordList = toJSON . map toCoords  -- this is just toJSON isn't it?
-
-      -- can we replace this infinite tower of toJSON calls with one toJSON call?
-      (ptype, cs) = case gType of
-        GeoPoint x y -> ("Point", toJSON [x, y])
-        GeoPoints coords -> ("MultiPoint", toCoords coords)
-        GeoLine coords -> ("LineString", toCoords coords)
-        GeoLines coords -> ("MultiLineString", toCoordList coords)
-        GeoPolygon coords -> ("Polygon", toCoordList coords)
-        GeoPolygons ccoords -> ("MultiPolygon", toJSON (map toCoordList ccoords))
-
-  in object [("type", ptype), ("coordinates", cs)]
-
-
-{-|
-
-Generate a data source that is a sphere for bounding global geographic data.
-The sphere will be subject to whatever projection is specified for the view.
-
-@
-'Graphics.Vega.VegaLite.toVegaLite'
-    [ 'sphere'
-    , 'projection' [ 'PrType' 'Orthographic' ]
-    , 'mark' 'Geoshape' [ 'MFill' "aliceblue" ]
-    ]
-@
-
-@since 0.4.0.0
--}
-sphere :: Data
-sphere = (VLData, object ["sphere" .= True])
-
-
-{-|
-
-Generate a grid of lines of longitude (meridians) and latitude
-(parallels).
-
-@
-let proj = 'projection' [ 'PrType' 'Orthographic' ]
-    sphereSpec = 'asSpec' [ 'sphere'
-                        , 'mark' 'Geoshape' [ 'MFill' "aliceblue" ] ]
-    gratSpec =
-        'asSpec'
-            [ 'graticule' [ 'GrStep' (5, 5) ]
-            , 'mark' 'Geoshape' [ 'MFilled' False, 'MStrokeWidth' 0.3 ]
-            ]
-in 'Graphics.Vega.VegaLite.toVegaLite' [ proj, 'layer' [ sphereSpec, gratSpec ] ]
-@
-
-@since 0.4.0.0
-
--}
-graticule ::
-  [GraticuleProperty] -- ^ An empty list uses the default parameters
-  -> Data
-graticule [] = (VLData, object ["graticule" .= True])
-graticule grProps =
-  (VLData, object ["graticule" .= object (map graticuleProperty grProps)])
-
-
-{-|
-
-Determine the properties of graticules. See the
-<https://vega.github.io/vega-lite/docs/data.html#graticule Vega-Lite documentation> for details.
-
-@since 0.4.0.0
-
--}
-data GraticuleProperty
-    = GrExtent (Double, Double) (Double, Double)
-    -- ^ Define the extent of both the major and minor graticules.
-    --   The range is given as longitude, latitude pairs of the
-    --   minimum and then maximum extent. The units are degrees.
-    | GrExtentMajor (Double, Double) (Double, Double)
-    -- ^ As @GrExtent@ but for the major graticule lines only.
-    | GrExtentMinor (Double, Double) (Double, Double)
-    -- ^ As @GrExtent@ but for the minor graticule lines only.
-    | GrStep (Double, Double)
-    -- ^ The step angles for the graticule lines, given as a longitude,
-    --   latitude pair defining the EW and NS intervals respectively.
-    --   The units are degrees.
-    --
-    --   By default major graticule lines extend to both poles but the
-    --   minor lines stop at ±80 degrees latitude.
-    | GrStepMajor (Double, Double)
-    -- ^ As @GrStep@ but for the major graticule lines only.
-    --
-    --   The default is @(90, 360)@.
-    | GrStepMinor (Double, Double)
-    -- ^ As @GrStep@ but for the minor graticule lines only.
-    --
-    --   The default is @(10, 10)@.
-    | GrPrecision Double
-    -- ^ The precision of the graticule. The units are degrees.
-    --   A smaller value reduces visual artifacts (steps) but takes longer
-    --   to render.
-    --
-    --   The default is @2.5@.
-
-
-graticuleProperty :: GraticuleProperty -> LabelledSpec
-graticuleProperty (GrExtent (lng1, lat1) (lng2, lat2)) =
-  "extent" .= [[lng1, lat1], [lng2, lat2]]
-graticuleProperty (GrExtentMajor (lng1, lat1) (lng2, lat2)) =
-  "extentMajor" .= [[lng1, lat1], [lng2, lat2]]
-graticuleProperty (GrExtentMinor (lng1, lat1) (lng2, lat2)) =
-  "extentMinor" .= [[lng1, lat1], [lng2, lat2]]
-graticuleProperty (GrStep (lng, lat)) = "step" .= [lng, lat]
-graticuleProperty (GrStepMajor (lng, lat)) = "stepMajor" .= [lng, lat]
-graticuleProperty (GrStepMinor (lng, lat)) = "stepMinor" .= [lng, lat]
-graticuleProperty (GrPrecision x) = "precision" .= x
 
 
 {-|
